[Chore][MoE] address Gemini review round 2 on PR #1514
- benchmarks/ops/bench_fused_moe_experts.py: allocate empty workspaces
  with explicit dtype to honor manifest's bf16/fp16 union (avoids
  validate-manifest dtype-fuzzer rejection)
- routed_expert/fused_routed_expert.py: use torch.mul(..., out=output) in
  the Padded forward when routed_scaling_factor != 1.0, mirroring the
  Nopad implementation and avoiding a temporary tensor allocation
- tests/test_fused_moe_experts.py: update the caplog logger name from
  the pre-refactor 'tileops.ops.moe.experts.nopad' to the new
  'tileops.ops.moe.routed_expert.fused_routed_expert' so the fallback
  warning is actually captured

+++ benchmarks/ops/bench_fused_moe_experts.py
@@ -155,8 +155,8 @@ def test_moe_experts_nopad_bench(
         hidden_size=hidden_size, ffn_size=ffn_size, dtype=dtype,
     )
     output = torch.empty(num_tokens, hidden_size, dtype=dtype, device="cuda")
-    ws1 = torch.empty(0, device="cuda")
-    ws2 = torch.empty(0, device="cuda")
+    ws1 = torch.empty(0, dtype=dtype, device="cuda")
+    ws2 = torch.empty(0, dtype=dtype, device="cuda")
 
     # -- TileOPs nopad (3WG persistent) --------------------------------------
     nopad = FusedMoEExpertsNopadPersistent3WGFwdOp(**kwargs)
